online-school-bot/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai>=1.12.0
python-dotenv>=1.0.0
slack-sdk>=3.27.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
line-bot-sdk>=3.5.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
numpy>=1.24.0
requests>=2.31.0
